Extract EXTENT constant from Bucket
Eliminates a circular require (bucket ⇢ circle_bucket ⇢ load_geometry ⇢ bucket).
