+++ js/render/painter.js
@@ -4,7 +4,7 @@ const browser = require('../util/browser');
 const mat4 = require('gl-matrix').mat4;
 const FrameHistory = require('./frame_history');
 const SourceCache = require('../source/source_cache');
-const EXTENT = require('../data/bucket').EXTENT;
+const EXTENT = require('../data/extent');
 const pixelsToTileUnits = require('../source/pixels_to_tile_units');
 const util = require('../util/util');
 const StructArrayType = require('../util/struct_array');
