+++ js/source/geojson_source.js
@@ -3,7 +3,7 @@
 const Evented = require('../util/evented');
 const util = require('../util/util');
 const window = require('../util/window');
-const EXTENT = require('../data/bucket').EXTENT;
+const EXTENT = require('../data/extent');
 
 /**
  * A source containing GeoJSON.
