+++ js/source/image_source.js
@@ -6,7 +6,7 @@ const LngLat = require('../geo/lng_lat');
 const Point = require('point-geometry');
 const Evented = require('../util/evented');
 const ajax = require('../util/ajax');
-const EXTENT = require('../data/bucket').EXTENT;
+const EXTENT = require('../data/extent');
 const RasterBoundsArray = require('../render/draw_raster').RasterBoundsArray;
 const Buffer = require('../data/buffer');
 const VertexArrayObject = require('../render/vertex_array_object');
